pyproject.toml
[tool.poetry]
name = "federated-engineers-portal-backend"
version = "0.1.0"
description = "A login portal for federated engineers to do their work"
authors = ["Ben van der Harg <ben.j.van.der.harg@postnl.nl>"]
readme = "README.md"
packages = [{include = "federated_engineers_portal_backend"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
python-dotenv = "^1.0.1"
toml = "^0.10.2"
uvicorn = "^0.29.0"
dependency-injector-fork = "^4.42.1"
boto3 = "^1.34.68"
dogpile-cache = "^1.3.2"
redis = "^5.0.3"
requests = "^2.31.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
jinja2 = "^3.1.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
